Build master deps from source

+++ ros1_build.sh
@@ -2,7 +2,7 @@
 set -e
 
 # install dependencies
-apt update && apt install -y lcov python3-pip libgtest-dev cmake && rosdep update
+apt update && apt install -y lcov python3-pip python-rosinstall libgtest-dev cmake && rosdep update
 cd /usr/src/gtest && cmake CMakeLists.txt && make && cp *.a /usr/lib
 apt update && apt install -y python3-colcon-common-extensions && pip3 install -U setuptools
 # nosetests needs coverage for Python 2
@@ -10,10 +10,22 @@ apt-get install python-pip -y && pip install -U coverage
 # enable Python coverage "https://github.com/ros/catkin/blob/kinetic-devel/cmake/test/nosetests.cmake#L59"
 export CATKIN_TEST_COVERAGE=1
 
-# use colcon as build tool to build the package, and optionally build tests
 . "/opt/ros/${ROS_DISTRO}/setup.sh"
+REPO_NAME=`echo $TRAVIS_BUILD_DIR | cut -c 33-`
 cd "/${ROS_DISTRO}_ws/"
-rosdep install --from-paths src --ignore-src --rosdistro "${ROS_DISTRO}" -r -y
+
+# use colcon as build tool to build the package, and optionally build tests
+if [ "$TRAVIS_BRANCH" == "master" ] && [ -f "./src/${REPO_NAME}/.rosinstall.master" ]; then
+    mkdir dep
+    cd "/${ROS_DISTRO}_ws/dep"
+    ln -s "../src/${REPO_NAME}/.rosinstall.master" .rosinstall
+    rosws update
+    cd "/${ROS_DISTRO}_ws/"
+    rosdep install --from-paths src dep --ignore-src --rosdistro "${ROS_DISTRO}" -r -y
+else
+    rosdep install --from-paths src --ignore-src --rosdistro "${ROS_DISTRO}" -r -y
+fi
+
 colcon build --cmake-args -DCMAKE_EXPORT_COMPILE_COMMANDS=ON -DCMAKE_CXX_FLAGS='-fprofile-arcs -ftest-coverage' -DCMAKE_C_FLAGS='-fprofile-arcs -ftest-coverage'
 if [ -z "${NO_TEST}" ];
 then
@@ -24,6 +36,12 @@ then
 
     # run unit tests
     . ./install/setup.sh
+
+    if [ "$TRAVIS_BRANCH" == "master" ];
+    then
+        touch dep/COLCON_IGNORE
+    fi
+
     colcon test
     colcon test-result --all
